+++ NEWS.md
@@ -1,4 +1,4 @@
-# semlrtp 0.0.1.1
+# semlrtp 0.0.1.3
 
 - First stable release. (0.0.1.0)
 - Revised `fix_to_zero()`, `lrt()`, and
@@ -10,3 +10,12 @@
   of the likelihood ratio tests
   encountered any error or warning.
   (0.0.1.1)
+- By default, `se = "bootstrap"` is used
+  in the original solution, it will
+  be forced to `se = "standard"` to
+  save processing time. This can be
+  turned off by the argument
+  `se_keep_bootstrap`. (0.0.1.2)
+- Updated `lrt()` and `lrtp()` to work
+  with models fitted by methods like
+  `MLR` and `MLM`. (0.0.1.3)